お知らせの有効期限機能と路線系統別お知らせ機能の導入に向けた管理画面の仕様変更等
FossilOrigin-Name: a658794a49e5dfd09873e2ea06101f09bcbf01a1ebcc5a74e106b39705512093

+++ unyohub/admin/announcements.php
@@ -20,7 +20,7 @@ function save_announcements ($json_path, $announcements) {
         $json_timestamp = time();
     }
     
-    if (fileowner($json_path) !== posix_geteuid()) {
+    if (file_exists($json_path) && fileowner($json_path) !== posix_geteuid()) {
         unlink($json_path);
     }
     
@@ -31,37 +31,71 @@ function save_announcements ($json_path, $announcements) {
 
 print_header();
 
-$json_path = "../config/announcements.json";
-$announcements = json_decode(file_get_contents($json_path), TRUE);
+if (empty($_GET["railroad_id"])) {
+    $railroad_id = "";
+    $json_path = "../config/announcements.json";
+} else {
+    $railroad_id = basename($_GET["railroad_id"]);
+    $json_path = "../data/".$railroad_id."/announcements.json";
+}
+
+if (file_exists($json_path)) {
+    $announcements = json_decode(file_get_contents($json_path), TRUE);
+} else {
+    $announcements = array();
+}
+
 
 $title_escaped = "";
 $content_html = "";
 
-if (isset($_POST["title"], $_POST["content"])) {
-    if (empty($_POST["title"]) || empty($_POST["content"])) {
-        print "<script> alert('【!】入力内容に空欄があります'); </script>";
-        
-        $title_escaped = addslashes($_POST["title"]);
-        $content_html = htmlspecialchars($_POST["content"]);
+$ts = time();
+
+$expiration_ts = $ts + 86400;
+$expiration_datetime = date("Y-m-d", $expiration_ts)."T".date("H:i", $expiration_ts);
+
+if (isset($_POST["title"], $_POST["content"], $_POST["expiration_datetime"])) {
+    $error_mes = "";
+    
+    if (empty($_POST["title"]) || empty($_POST["content"]) || empty($_POST["expiration_datetime"])) {
+        $error_mes = "入力内容に空欄があります";
     } elseif (isset($_POST["one_time_token"]) && $user->check_one_time_token($_POST["one_time_token"])) {
-        $last_modified_timestamp = time();
+        $expiration_timestamp = strtotime($_POST["expiration_datetime"]);
         
-        array_unshift($announcements, array(
-            "title" => $_POST["title"],
-            "is_important" => !empty($_POST["is_important"]),
-            "content" => $_POST["content"],
-            "user_id" => $user->get_id(),
-            "last_modified_timestamp" => $last_modified_timestamp
-        ));
-        
-        save_announcements($json_path, $announcements);
-        
-        print "<script> alert('お知らせを追加しました'); </script>";
+        if ($expiration_timestamp > $ts + 600) {
+            $last_modified_timestamp = time();
+            
+            for ($cnt = 0; isset($announcements[$cnt]); $cnt++) {
+                if ($announcements[$cnt]["expiration_timestamp"] < $ts) {
+                    array_splice($announcements, $cnt, 1);
+                }
+            }
+            
+            array_unshift($announcements, array(
+                "title" => $_POST["title"],
+                "is_important" => !empty($_POST["is_important"]),
+                "content" => $_POST["content"],
+                "user_id" => $user->get_id(),
+                "expiration_timestamp" => $expiration_timestamp,
+                "last_modified_timestamp" => $last_modified_timestamp
+            ));
+            
+            save_announcements($json_path, $announcements);
+            
+            print "<script> alert('お知らせを追加しました'); </script>";
+        } else {
+            $error_mes = "有効期限が不十分です";
+        }
     } else {
-        print "<script> alert('【!】ワンタイムトークンが無効です。再送信してください。'); </script>";
+        $error_mes = "ワンタイムトークンが無効です。再送信してください。";
+    }
+    
+    if (!empty($error_mes)) {
+        print "<script> alert('【!】".$error_mes."'); </script>";
         
         $title_escaped = addslashes($_POST["title"]);
         $content_html = htmlspecialchars($_POST["content"]);
+        $expiration_datetime = addslashes($_POST["expiration_datetime"]);
     }
 } elseif (isset($_POST["delete_index"])) {
     if (isset($_POST["one_time_token"]) && $user->check_one_time_token($_POST["one_time_token"])) {
@@ -78,11 +112,37 @@ if (isset($_POST["title"], $_POST["content"])) {
 print "<article>";
 print "<h2>お知らせの編集</h2>";
 
+print "<h3>対称の路線系統</h3>";
+print "<select class='wide_select' onchange='location.href = location.origin + location.pathname + \"?railroad_id=\" + this.value;'>";
+print "<option value=''>全体のお知らせ</option>";
+
+$railroad_list = file("../config/railroads.txt", FILE_IGNORE_NEW_LINES | FILE_SKIP_EMPTY_LINES);
+
+foreach ($railroad_list as $railroad) {
+    $railroad = trim($railroad);
+    if (empty($railroad)) {
+        continue;
+    }
+    
+    $railroad_info = @json_decode(@file_get_contents("../data/".$railroad."/railroad_info.json"), TRUE);
+    if (empty($railroad_info)) {
+        continue;
+    }
+    
+    print "<option value='".addslashes($railroad)."'";
+    if ($railroad === $railroad_id) {
+        print " selected='selected'";
+    }
+    print ">".htmlspecialchars($railroad_info["railroad_name"])."</option>";
+}
+
+print "</select><br>";
+
 $token_html = "<input type='hidden' name='one_time_token' value='".$user->create_one_time_token()."'>";
 
 print "<input type='checkbox' id='new_announcement'><label for='new_announcement' class='drop_down'>【+】新しいお知らせを追加</label>";
 
-print "<div><form action='announcements.php' method='post'>";
+print "<div><form action='announcements.php?railroad_id=".$railroad_id."' method='post'>";
 print $token_html;
 
 print "<h3>件名</h3>";
@@ -91,12 +151,15 @@ print "<input type='text' name='title' value='".$title_escaped."'>";
 print "<h3>本文</h3>";
 print "<textarea name='content'>".$content_html."</textarea>";
 
+print "<h3>有効期限</h3>";
+print "<input type='datetime-local' name='expiration_datetime' value='".$expiration_datetime."'><br>";
+
 print "<input type='checkbox' name='is_important' id='is_important' class='toggle' value='YES'><label for='is_important'>重要なお知らせ</label><br>";
 
 print "<button type='submit' class='wide_button'>追加</button>";
 print "</form></div>";
 
-print "<form action='announcements.php' method='post' id='delete_form'>";
+print "<form action='announcements.php?railroad_id=".$railroad_id."' method='post' id='delete_form'>";
 print $token_html;
 print "<input type='hidden' name='delete_index' id='delete_index'>";
 print "</form>";
@@ -113,6 +176,10 @@ EOM;
 
 print "<div id='announcements_area'>";
 for ($cnt = 0; isset($announcements[$cnt]); $cnt++) {
+    if ($announcements[$cnt]["expiration_timestamp"] < $ts) {
+        continue;
+    }
+    
     $user = $wakarana->get_user($announcements[$cnt]["user_id"]);
     if (is_object($user)) {
         $user_name = $user->get_name();
@@ -127,7 +194,11 @@ for ($cnt = 0; isset($announcements[$cnt]); $cnt++) {
     print "'>".htmlspecialchars($announcements[$cnt]["title"])."</label>";
     print "<div><div class='announcement'>";
     print nl2br(htmlspecialchars($announcements[$cnt]["content"]));
-    print "<small>".htmlspecialchars($user_name)."　".date("Y/n/j G:i", $announcements[$cnt]["last_modified_timestamp"])."</small>";
+    print "<small>".htmlspecialchars($user_name)."　".date("Y/n/j G:i", $announcements[$cnt]["last_modified_timestamp"])." (残り";
+    if ($announcements[$cnt]["expiration_timestamp"] >= $ts + 86400) {
+        print " ".floor(($announcements[$cnt]["expiration_timestamp"] - $ts) / 86400)."日";
+    }
+    print " ".(floor(($announcements[$cnt]["expiration_timestamp"] - $ts) / 3600) % 24)."時間 有効)</small>";
     print "</div><button type='submit' class='wide_button' onclick='delete_announcement(\"".$cnt."\");'>このお知らせを削除</button>";
     print "</div>";
 }
